refactored tests and changed name from master to default Branch
